+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     - `DeleteMessagesFromParticipant`: delete list of messages for a participant
     - `GetResearcherMessages`: get all researcher messages form all studies
     - `DeleteResearcherMessages`: remove researcher messages with study key and id.
+    - `RunRulesForSingleParticipant`: run custom rules to a single participant.
 
 - Participant File Upload endpoint: method to upload files for study participants
 - Can configure root path of the persistent storage with PERSISTENCE_STORE_ROOT_PATH env variable.